Add PNG versions of application icons (in various fixed sizes)
This is an experimental commit that makes KeePassXC use fixed-size
PNG icons instead of relying on Qt's SVG scaling. Note that the
result is not completely satisfactory (and may be system-dependent).

It comes with a shell script (utils/makepng.sh) that converts the
SVG files to PNG. The script contains three implementations of the
SVG->PNG conversion routine, two with ImageMagick and one with
Inkscape. Instructions for use are in the script's header comment.

Fixes #475/#4066 (after some more work).